LayoutTests/media/progress-event-expected.txt

EVENT(loadstart) TEST(!event.lengthComputable && event.loaded==0 && event.total==0) OK
EVENT(load) TEST(event.lengthComputable && event.loaded>0 && event.total>0) OK
END OF TEST
